better file arrangment and code alignment
